Move testutils where they truly belong: in tests
